% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_operating_chars.R
\name{mcse_prop}
\alias{mcse_prop}
\title{Monte-Carlo standard error of a proportion}
\usage{
mcse_prop(p, n)
}
\arguments{
\item{p}{Estimated proportion.}

\item{n}{Number of replicates.}
}
\value{
The binomial MCSE \code{sqrt(p (1 - p) / n)}.
}
\description{
Monte-Carlo standard error of a proportion
}
\keyword{internal}
