% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_oc_bars}
\alias{plot_oc_bars}
\title{Operating characteristics as a labelled bar chart with MCSE error bars}
\usage{
plot_oc_bars(oc)
}
\arguments{
\item{oc}{An operating-characteristics data frame from \code{\link[=summarise_oc]{summarise_oc()}}.}
}
\value{
A ggplot object showing the proportion-type measures.
}
\description{
Operating characteristics as a labelled bar chart with MCSE error bars
}
\examples{
res <- run_trial_simulation(0.3, 0.45, n_max = 150, n_interim = 75,
                            n_sims = 300)
plot_oc_bars(res$oc)
}
