% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.}

\item{options}{A named list of options passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{enableBookmarking}{Bookmarking setting passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the request.}

\item{...}{arguments to pass to golem_opts. See \code{?golem::get_golem_options}.}
}
\value{
A Shiny application object (invisibly when run interactively).
}
\description{
Launches the bayesadapt Bayesian adaptive clinical-trial simulator.
}
