% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{theme_bayesadapt}
\alias{theme_bayesadapt}
\title{Minimal ggplot theme for operating-characteristic figures}
\usage{
theme_bayesadapt()
}
\description{
Minimal ggplot theme for operating-characteristic figures
}
\keyword{internal}
